+++ App/cli/main.swift
@@ -2,4 +2,4 @@ import Foundation
 import xcbuddykit
 
 var registry = CommandRegistry()
-registry.run()
+try registry.run()