+++ App/xcbuddykit/Sources/Commands/CommandRegistry.swift
@@ -10,6 +10,9 @@ public final class CommandRegistry {
     /// Printer.
     private let printer: Printing
 
+    /// Error handler.
+    private let errorHandler: ErrorHandling
+
     // Registered commands.
     var commands: [Command] = []
 
@@ -19,6 +22,7 @@ public final class CommandRegistry {
     /// Initializes the command registry
     public init(processArguments: @escaping () -> [String] = CommandRegistry.processArguments) {
         printer = Printer()
+        errorHandler = ErrorHandler(printer: printer)
         parser = ArgumentParser(usage: "<command> <options>",
                                 overview: "Your Xcode buddy")
         self.processArguments = processArguments
@@ -43,19 +47,16 @@ public final class CommandRegistry {
     }
 
     /// Runs the command line interface.
-    public func run() {
+    public func run() throws {
+        let parsedArguments = try parse()
         do {
-            let parsedArguments = try parse()
             try process(arguments: parsedArguments)
-        } catch let error as ArgumentParserError {
-            printer.print(errorMessage: error.localizedDescription)
-            exit(1)
-        } catch let error as CustomStringConvertible {
-            printer.print(errorMessage: error.description)
-            exit(1)
         } catch {
-            printer.print(errorMessage: error.localizedDescription)
-            exit(1)
+            // Errors shouldn't reach this point.
+            // In case we do we should check where the error is coming from and:
+            //  - Handle them if we can.
+            //  - Use the error handler from the context to fail the execution.
+            errorHandler.fatal(error: FatalError.bugSilent(error))
         }
     }
 
@@ -78,6 +79,6 @@ public final class CommandRegistry {
             parser.printUsage(on: stdoutStream)
             return
         }
-        try command.run(with: arguments)
+        command.run(with: arguments)
     }
 }