+++ App/xcbuddykit/Sources/Commands/GenerateCommand.swift
@@ -13,12 +13,23 @@ enum GenerateCommandError: Error, CustomStringConvertible, Equatable {
 }
 
 public class GenerateCommand: NSObject, Command {
+    /// Command name.
     public let command = "generate"
+
+    /// Command description.
     public let overview = "Generates an Xcode workspace to start working on the project."
+
+    /// Graph loader context.
     fileprivate let graphLoaderContext: GraphLoaderContexting
-    fileprivate let commandsContext: CommandsContexting
+
+    /// Graph loader.
     fileprivate let graphLoader: GraphLoading
+
+    /// Workspace generator.
     fileprivate let workspaceGenerator: WorkspaceGenerating
+
+    /// Context.
+    fileprivate let context: CommandsContexting
 
     /// Path argument.
     let pathArgument: OptionArgument<String>
@@ -28,30 +39,30 @@ public class GenerateCommand: NSObject, Command {
     /// - Parameter parser: argument parser.
     public required convenience init(parser: ArgumentParser) {
         self.init(graphLoaderContext: GraphLoaderContext(),
-                  commandsContext: CommandsContext(),
                   graphLoader: GraphLoader(),
                   workspaceGenerator: WorkspaceGenerator(),
-                  parser: parser)
+                  parser: parser,
+                  context: CommandsContext())
     }
 
     /// Initializes the command with the printer and the graph loading context.
     ///
     /// - Parameters:
     ///   - graphLoaderContext: graph loading context.
-    ///   - commandsContext: commands context.
     ///   - graphLoader: graph loader.
     ///   - workspaceGenerator: workspace generator,
     ///   - parser: argument parser.
+    ///   - context: context.
     init(graphLoaderContext: GraphLoaderContexting,
-         commandsContext: CommandsContexting,
          graphLoader: GraphLoading,
          workspaceGenerator: WorkspaceGenerating,
-         parser: ArgumentParser) {
+         parser: ArgumentParser,
+         context: CommandsContexting) {
         let subParser = parser.add(subparser: command, overview: overview)
         self.graphLoaderContext = graphLoaderContext
-        self.commandsContext = commandsContext
         self.graphLoader = graphLoader
         self.workspaceGenerator = workspaceGenerator
+        self.context = context
         pathArgument = subParser.add(option: "--path",
                                      shortName: "-p",
                                      kind: String.self,
@@ -63,13 +74,15 @@ public class GenerateCommand: NSObject, Command {
     ///
     /// - Parameter _: argument parser arguments.
     /// - Throws: an error if the command cannot be executed.
-    public func run(with arguments: ArgumentParser.Result) throws {
-        var path: AbsolutePath! = arguments.get(pathArgument).map({ AbsolutePath($0) })
-        if path == nil {
-            path = AbsolutePath.current
+    public func run(with arguments: ArgumentParser.Result) {
+        context.errorHandler.try {
+            var path: AbsolutePath! = arguments.get(pathArgument).map({ AbsolutePath($0) })
+            if path == nil {
+                path = AbsolutePath.current
+            }
+            let context = try GeneratorContext(graph: graphLoader.load(path: path))
+            try workspaceGenerator.generate(path: path, context: context)
+            self.context.printer.print(section: "Generate command succeeded 🎉")
         }
-        let context = try GeneratorContext(graph: graphLoader.load(path: path))
-        try workspaceGenerator.generate(path: path, context: context)
-        context.printer.print(section: "Generate command succeeded 🎉")
     }
 }