+++ App/xcbuddykit/Sources/Generator/GeneratorError.swift
@@ -1,4 +1,7 @@
 import Foundation
 
-enum GeneratorError: Error {
+enum GeneratorError: Error, CustomStringConvertible {
+    var description: String {
+        return ""
+    }
 }